Move sim_debug to the BIOS directory
