+++ litex/soc/software/bios/sim_debug.c
