+++ litex/soc/software/bios/sim_debug.h
